Removing of "/" not required
Removing of the trailing "/" is not required as the app will always be installed on root i.e. subdomian.domain.tld
So the path will always remain "/" . 
Removing "/" will make the "location  {...code} "when it should be "location / {..code}" in nginx.conf.
Thats why if "/"is removed every time a upgrade is performed it results in error of nginx.